Add support others posix
If you have an iPhone configured with Spanish language. When you get the data from the Bing service, they do not show the regions from the country where I am. For example, when I change the posix of the United States for one in Spanish, in that case if all the regions are shown. If I change the language of my phone